requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# HTTP client for API calls
aiohttp>=3.8.0

# Logging and utilities
python-json-logger>=2.0.0

# Optional: For testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 

# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0